refactor swift ui
