+++ apps/macos/LauncherApp/look-app/Views/Commands/CalcCommand.swift
