+++ apps/macos/LauncherApp/look-app/Views/Commands/KillCommand.swift
