+++ apps/macos/LauncherApp/look-app/Views/Commands/ShellCommand.swift
