+++ apps/macos/LauncherApp/look-app/Views/Commands/SystemInfoCommand.swift
