+++ apps/macos/LauncherApp/look-app/Views/Launcher/LauncherBannerView.swift
