+++ apps/macos/LauncherApp/look-app/Views/Launcher/LauncherRowView.swift
