+++ apps/macos/LauncherApp/look-app/Views/Launcher/LookupDefinitionPanelView.swift
