+++ apps/macos/LauncherApp/look-app/Views/Launcher/ResultPreviewView.swift
